requirements.txt
nest-asyncio
python-telegram-bot
PyStarDict